created inventories to be static rather than associated with ScrollCommand
