+++ src/main/java/apple/ScrollMain.java
@@ -9,6 +9,7 @@ import org.bukkit.plugin.java.JavaPlugin;
 public class ScrollMain extends JavaPlugin {
     @Override
     public void onEnable() {
+        new ScrollInventories(this);
         new CreateCommand(this);
         new DuplicateCommand(this);
         new ClickListener(this);
