+++ src/main/java/apple/commands/ScrollCommand.java
@@ -1,6 +1,7 @@
 package apple.commands;
 
 import apple.InventoryChest;
+import apple.ScrollInventories;
 import apple.ScrollMain;
 import org.bukkit.Bukkit;
 import org.bukkit.Material;
@@ -23,12 +24,9 @@ import java.util.List;
 
 public class ScrollCommand implements CommandExecutor {
     private JavaPlugin plugin;
-    private Inventory scrollInvAll;
 
     public ScrollCommand(ScrollMain plugin) {
         this.plugin = plugin;
-        update();
-
         // set the scroll command to execute here
         PluginCommand command = plugin.getCommand("scroll");
         if (command == null) {
@@ -37,44 +35,6 @@ public class ScrollCommand implements CommandExecutor {
         command.setExecutor(this);
     }
 
-    public void update() {
-        scrollInvAll = Bukkit.createInventory(new InventoryChest(plugin, 54, "Scrolls"), InventoryType.CHEST);
-        File file = new File(plugin.getDataFolder() + File.separator + "scrollInv" + File.separator + "scrollInv.yml");
-        YamlConfiguration config = YamlConfiguration.loadConfiguration(file);
-        ConfigurationSection configInv = config.getConfigurationSection("inventory");
-        ConfigurationSection configInvAll = configInv.getConfigurationSection("all");
-        int i = 1;
-        ConfigurationSection configInvAllItem = configInvAll.getConfigurationSection(String.format("item%d", i++));
-
-        // get all the items in inv all
-        while (configInvAllItem != null) {
-            ItemStack item = getItemFromConfig(configInvAllItem);
-            configInvAllItem = configInvAll.getConfigurationSection(String.format("item%d", i++));
-            scrollInvAll.addItem(item);
-        }
-    }
-
-    private static ItemStack getItemFromConfig(ConfigurationSection config) {
-        //todo update these vvv
-        String type = config.getString("material");
-        ItemStack item = new ItemStack(Material.getMaterial(type));
-        List<String> lore = new ArrayList<String>(4);
-        ConfigurationSection configLore = config.getConfigurationSection("lore");
-        // get the lore
-        int i = 1;
-        String loreLine = configLore.getString(String.format("line%d", i++));
-        while (loreLine != null) {
-            lore.add(loreLine);
-            loreLine = configLore.getString(String.format("line%d", i++));
-        }
-        ItemMeta im = item.getItemMeta();
-        im.setLore(lore);
-        item.setItemMeta(im);
-        item.setAmount(1);
-
-        // 1 item in the stack
-        return item;
-    }
 
     @Override
     public boolean onCommand(CommandSender commandSender, Command command, String s, String[] strings) {
@@ -86,7 +46,7 @@ public class ScrollCommand implements CommandExecutor {
             return false;
 
         // open the scroll inventory
-        player.openInventory(scrollInvAll);
+        player.openInventory(ScrollInventories.scrollInvAll);
 
         return false;
     }
